Moved recast methods to the Recast class
