+++ src/main/java/meteordevelopment/meteorclient/systems/modules/movement/elytrafly/modes/Recast.java
@@ -14,7 +14,13 @@ import meteordevelopment.meteorclient.systems.modules.movement.elytrafly.ElytraF
 import meteordevelopment.meteorclient.systems.modules.movement.elytrafly.ElytraFlightModes;
 import meteordevelopment.meteorclient.systems.modules.movement.elytrafly.ElytraFly;
 import meteordevelopment.meteorclient.utils.misc.input.Input;
+import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.option.KeyBinding;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.effect.StatusEffects;
+import net.minecraft.item.ElytraItem;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
 import net.minecraft.network.packet.c2s.play.ClientCommandC2SPacket;
 import net.minecraft.network.packet.s2c.play.PlayerPositionLookS2CPacket;
 
@@ -33,7 +39,7 @@ public class Recast extends ElytraFlightMode {
         super.onTick();
 
         // Make sure all the conditions are met (player has an elytra, isn't in water, etc)
-        if (ElytraFly.checkConditions(mc.player)) {
+        if (checkConditions(mc.player)) {
 
             if (elytraFly.flightMode.get() == ElytraFlightModes.Recast) {
                 mc.player.setSprinting(true);
@@ -73,6 +79,28 @@ public class Recast extends ElytraFlightMode {
     private void setPressed(KeyBinding key, boolean pressed) {
         key.setPressed(pressed);
         Input.setKeyState(key, pressed);
+    }
+
+    public static boolean recastElytra(ClientPlayerEntity player) {
+        if (checkConditions(player) && ignoreGround(player)) {
+            player.networkHandler.sendPacket(new ClientCommandC2SPacket(player, ClientCommandC2SPacket.Mode.START_FALL_FLYING));
+            return true;
+        } else return false;
+    }
+
+    public static boolean checkConditions(ClientPlayerEntity player) {
+        ItemStack itemStack = player.getEquippedStack(EquipmentSlot.CHEST);
+        return (!player.getAbilities().flying && !player.hasVehicle() && !player.isClimbing() && itemStack.isOf(Items.ELYTRA) && ElytraItem.isUsable(itemStack));
+    }
+
+    private static boolean ignoreGround(ClientPlayerEntity player) {
+        if (!player.isTouchingWater() && !player.hasStatusEffect(StatusEffects.LEVITATION)) {
+            ItemStack itemStack = player.getEquippedStack(EquipmentSlot.CHEST);
+            if (itemStack.isOf(Items.ELYTRA) && ElytraItem.isUsable(itemStack)) {
+                player.startFallFlying();
+                return true;
+            } else return false;
+        } else return false;
     }
 
     private float getSmartYawDirection() {
